+++ src/routes/+layout.svelte
@@ -1,12 +1,22 @@
 <script lang="ts">
+	import { QueryClientProvider } from '@tanstack/svelte-query';
+	import { SvelteQueryDevtools } from '@tanstack/svelte-query-devtools';
+	import { dev } from '$app/environment';
 	import favicon from '$lib/assets/favicon.svg';
+	import type { LayoutData } from './$types';
 
-	let { children } = $props();
-	import "../app.css";
+	let { children, data }: { children: import('svelte').Snippet; data: LayoutData } = $props();
+
+	import '../app.css';
 </script>
 
 <svelte:head>
 	<link rel="icon" href={favicon} />
 </svelte:head>
 
-{@render children()}
+<QueryClientProvider client={data.queryClient}>
+	{@render children()}
+	{#if dev}
+		<SvelteQueryDevtools initialIsOpen={false} buttonPosition="bottom-right" />
+	{/if}
+</QueryClientProvider>